Refactor: cts: Move cts.watcher to _cts module.
